refactor(codemode): collapse runner types and fold the Promise global into promises
