+++ packages/codemode/src/interpreter/globals.ts
@@ -7,7 +7,6 @@ import { jsonGlobal } from "../stdlib/json.js"
 import { mathGlobal } from "../stdlib/math.js"
 import { numberGlobal } from "../stdlib/number.js"
 import { objectGlobal } from "../stdlib/object.js"
-import { promiseGlobal } from "../stdlib/promise.js"
 import { regexpGlobal } from "../stdlib/regexp.js"
 import { stringGlobal } from "../stdlib/string.js"
 import { uriGlobal, urlGlobal, urlSearchParamsGlobal } from "../stdlib/url.js"
@@ -16,7 +15,7 @@ import { ToolReference } from "../tool-runtime.js"
 import { errorGlobal } from "./errors.js"
 import { HostFunction } from "./host.js"
 import { AsyncIteratorSymbol, InterpreterRuntimeError, IteratorSymbol } from "./model.js"
-import type { PromiseRuntime } from "./promises.js"
+import { promiseGlobal, type PromiseRuntime } from "./promises.js"
 import type { Runner } from "./runner.js"
 
 /** What the built-in globals need from the interpreter that owns them. */
